requirements.txt
flask==3.0.0
pdfplumber==0.10.2
python-docx==0.8.11
spacy==3.7.2
reportlab==4.0.8
python-magic==0.4.27
transformers==4.36.0